Add files via upload
Car Rental System
This is a project for Car Rental System in c++. A user can perform many kinds of management tasks to manage rent and car records like adding a new car as well as modifying and removing it. Also, users can view all rental records, check car details as well as rent and change rent details from rental records. The system automates the basic car rental functions to aid in the day-to-day operations of a car rental staff. Also, the goal of the system is to allow for storing details of a large number of car details and allow for add, retrieve and update details of any car. So, the car rental system using c++ utilizes JsonCpp library that allows manipulating JSON values.

The basic features
Add Car: This operation is necessary when user adds a new car. This option takes inputs of car details.
Update Car Details: This function updates details of car in the car records.
Remove Car: This operation clears the existing record of a car from the records. So, be careful while performing this operation because the system could loose any important data.
List all cars: Show all the available cars from the record of car rental system.
Check Car
Rent a Car
Modify Rent Records
The record of cars and rents in the car rental system is stored in the cars.json and rents.json file respectively. Moreover, the system records renter name, car number, id and rent price.

This system is a small project that implements only some of the complex features of car rental service. However, one can use this project to make complex system by adding more details/properties to the records as well as features and functions to the system.